+++ cloud/log-server/src/main/kotlin/pers/acp/admin/log/conf/LogServerCustomerConfiguration.kt
@@ -1,10 +1,8 @@
 package pers.acp.admin.log.conf
 
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.cloud.context.config.annotation.RefreshScope
 import org.springframework.stereotype.Component
-import pers.acp.spring.cloud.component.CloudTools
 
 /**
  * @author zhang by 30/01/2019
@@ -12,12 +10,7 @@ import pers.acp.spring.cloud.component.CloudTools
  */
 @Component
 @RefreshScope
-class LogServerCustomerConfiguration @Autowired constructor(cloudTools: CloudTools) {
-
-    var serverIp: String = cloudTools.getServerIp()
-
-    var serverPort: Int = cloudTools.getServerPort()
-
+class LogServerCustomerConfiguration {
     @Value("\${logging.file.name}")
     var logFile: String = ""