+++ common/deploy-server/src/main/kotlin/pers/acp/admin/deploy/bus/listener/DeployExecuteEventListener.kt
@@ -7,21 +7,23 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.ApplicationListener
 import org.springframework.stereotype.Component
 import pers.acp.admin.common.event.ExecuteBusEvent
-import pers.acp.admin.deploy.conf.DeployServerCustomerConfiguration
 import pers.acp.admin.deploy.constant.DeployConstant
 import pers.acp.admin.deploy.domain.DeployTaskDomain
 import pers.acp.core.task.BaseAsyncTask
 import pers.acp.core.task.threadpool.ThreadPoolService
 import pers.acp.spring.boot.interfaces.LogAdapter
+import pers.acp.spring.cloud.component.CloudTools
 import pers.acp.spring.cloud.lock.DistributedLock
 
 @Component
 class DeployExecuteEventListener @Autowired
-constructor(private val logAdapter: LogAdapter,
-            private val distributedLock: DistributedLock,
-            private val objectMapper: ObjectMapper,
-            private val deployServerCustomerConfiguration: DeployServerCustomerConfiguration,
-            private val deployTaskDomain: DeployTaskDomain) : ApplicationListener<ExecuteBusEvent> {
+constructor(
+    private val logAdapter: LogAdapter,
+    private val cloudTools: CloudTools,
+    private val distributedLock: DistributedLock,
+    private val objectMapper: ObjectMapper,
+    private val deployTaskDomain: DeployTaskDomain
+) : ApplicationListener<ExecuteBusEvent> {
     override fun onApplicationEvent(executeBusEvent: ExecuteBusEvent) {
         if (executeBusEvent.message == DeployConstant.busMessageExecuteDeploy) {
             logAdapter.info("收到执行部署任务事件")
@@ -29,30 +31,30 @@ constructor(private val logAdapter: LogAdapter,
                 logAdapter.debug(objectMapper.writeValueAsString(executeBusEvent))
                 if (executeBusEvent.paramList.isNotEmpty()) {
                     ThreadPoolService.getInstance(1, 1, Int.MAX_VALUE, DeployConstant.busMessageExecuteDeploy)
-                            .addTask(object : BaseAsyncTask(DeployConstant.busMessageExecuteDeploy, false) {
-                                override fun beforeExecuteFun(): Boolean = true
-                                override fun executeFun(): Any? {
-                                    val deployTaskId = executeBusEvent.paramList.first()
-                                    try {
-                                        if (getLock()) {
-                                            try {
-                                                logAdapter.info("开始执行部署任务...")
-                                                runBlocking {
-                                                    deployTaskDomain.doExecuteTask(deployTaskId)
-                                                }
-                                                logAdapter.info("部署任务执行完成！")
-                                            } finally {
-                                                releaseLock()
+                        .addTask(object : BaseAsyncTask(DeployConstant.busMessageExecuteDeploy, false) {
+                            override fun beforeExecuteFun(): Boolean = true
+                            override fun executeFun(): Any? {
+                                val deployTaskId = executeBusEvent.paramList.first()
+                                try {
+                                    if (getLock()) {
+                                        try {
+                                            logAdapter.info("开始执行部署任务...")
+                                            runBlocking {
+                                                deployTaskDomain.doExecuteTask(deployTaskId)
                                             }
+                                            logAdapter.info("部署任务执行完成！")
+                                        } finally {
+                                            releaseLock()
                                         }
-                                    } catch (e: Exception) {
-                                        logAdapter.error(e.message, e)
                                     }
-                                    return true
+                                } catch (e: Exception) {
+                                    logAdapter.error(e.message, e)
                                 }
+                                return true
+                            }
 
-                                override fun afterExecuteFun(result: Any) {}
-                            })
+                            override fun afterExecuteFun(result: Any) {}
+                        })
                     runBlocking {
                         delay(DeployConstant.DISTRIBUTED_LOCK_TIME_OUT)
                     }
@@ -64,15 +66,19 @@ constructor(private val logAdapter: LogAdapter,
     }
 
     private fun getLock(): Boolean =
-            distributedLock.getLock(DeployConstant.DEPLOY_EXECUTE_DISTRIBUTED_LOCK_PREFIX + deployServerCustomerConfiguration.serverIp,
-                    deployServerCustomerConfiguration.serverIp + ":" + deployServerCustomerConfiguration.serverPort,
-                    DeployConstant.DISTRIBUTED_LOCK_TIME_OUT)
+        distributedLock.getLock(
+            DeployConstant.DEPLOY_EXECUTE_DISTRIBUTED_LOCK_PREFIX + cloudTools.getServerIp(),
+            cloudTools.getServerIp() + ":" + cloudTools.getServerPort(),
+            DeployConstant.DISTRIBUTED_LOCK_TIME_OUT
+        )
 
     private fun releaseLock() {
         runBlocking {
             delay(DeployConstant.DISTRIBUTED_LOCK_TIME_OUT)
         }
-        distributedLock.releaseLock(DeployConstant.DEPLOY_EXECUTE_DISTRIBUTED_LOCK_PREFIX + deployServerCustomerConfiguration.serverIp,
-                deployServerCustomerConfiguration.serverIp + ":" + deployServerCustomerConfiguration.serverPort)
+        distributedLock.releaseLock(
+            DeployConstant.DEPLOY_EXECUTE_DISTRIBUTED_LOCK_PREFIX + cloudTools.getServerIp(),
+            cloudTools.getServerIp() + ":" + cloudTools.getServerPort()
+        )
     }
 }