+++ doc/version_history.md
@@ -16,6 +16,7 @@
 > - oauth-server
 >   - [Upgrade] 取消静态ip配置 server.address
 >   - [Upgrade] 使用动态端口
+>   - [Upgrade] 修复机构信息查询时，vo对象缺失area值的问题
 > - route-server
 >   - [Upgrade] 取消静态ip配置 server.address
 >   - [Upgrade] 使用动态端口